Milestone reached!
I've created a genetic bot (genbot1) as well as a control bot (genbotcontrol) and updated game_runner.py so that I can run games over several generations of the genetic bot.
My testing using randombot, and comparing the results of genbot1 with the results of genbotcontrol, suggests that some level of improvement has been achieved with genbot1.
The next step is to improve the performance of the system, and then try running with more samples, and over more generations.